One of these BIG commits you shouldn't do. Added the frontend in a separate directory with a separate build and made everythhing work together for the first time.

+++ sut/src/main/groovy/se/tarnowski/agdp2013/invoice/InvoiceRepository.groovy
@@ -6,7 +6,7 @@ interface InvoiceRepository {
 
     void updateInvoice(Invoice invoice)
 
-    Invoice findByInvoiceNumber(long invoiceNumber)
+    Invoice findByInvoiceNumber(BigInteger invoiceNumber)
 
 
 }
